removed UserSongCard component for simplicity sake, working on PUT and GET functionality in user dashboard
